sound now works
